Extract middleware test to utils
